fix: address devedup-rl review — split plaintext suite out of oversized processor spec (ROK-1354)
discord-notification.processor.spec.ts hit the 750-counted-line max-lines
error (779) after the ROK-1354 classification blocks landed. Moved the
buildPlaintextContent describe block (ROK-756/822/918, pure-function tests
of a format-helpers export) into format-helpers.spec.ts where it belongs.
No test changes — 132 tests across the 3 affected suites still pass.
